// src/api/generated/models/GetUserProfile.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * User profile response
 */
export type GetUserProfile = {
    /**
     * Profile image URL
     */
    profile_image_url?: (string | null);
    /**
     * User alias/nickname
     */
    aka?: (string | null);
    /**
     * Sui blockchain address
     */
    sui_address?: (string | null);
};